Reorganized documents tests
