+++ test/config/jest.config.storage.js
@@ -1,5 +1,5 @@
 module.exports = {
   ...require('./jest.config'),
-  testRegex: ['/test/functional-api/upload/.*\\.it-spec\\.ts'],
+  testRegex: ['/test/functional-api/storage/.*\\.it-spec\\.ts'],
   coverageDirectory: '<rootDir>/coverage-ci',
 };
