+++ test/functional-api/storage/auth/public-space-document-auth.it-spec.ts
@@ -8,15 +8,15 @@ import {
   uploadFileOnRef,
   uploadFileOnStorageBucket,
   uploadImageOnVisual,
-} from './upload.params';
+} from '../upload.params';
 import path from 'path';
-import { deleteOrganizationCodegen } from '../organization/organization.request.params';
+import { deleteOrganizationCodegen } from '../../organization/organization.request.params';
 import { createOrgAndSpaceWithUsersCodegen } from '@test/utils/data-setup/entities';
-import { lookupProfileVisuals } from '../lookup/lookup-request.params';
+import { lookupProfileVisuals } from '../../lookup/lookup-request.params';
 import {
   deleteSpaceCodegen,
   updateSpaceSettingsCodegen,
-} from '../journey/space/space.request.params';
+} from '../../journey/space/space.request.params';
 import {
   sorted__create_read_update_delete_grant,
   sorted__create_read_update_delete_grant_contribute,
@@ -28,14 +28,14 @@ import {
 import {
   createLinkCollectionCalloutCodegen,
   createLinkOnCalloutCodegen,
-} from '../callout/collection-of-links/collection-of-links-callout.params.request';
+} from '../../callout/collection-of-links/collection-of-links-callout.params.request';
 import {
   calloutLinkContributionStorageConfigCodegen,
   calloutPostCardStorageConfigCodegen,
   calloutStorageConfigCodegen,
   calloutWhiteboardStorageConfigCodegen,
   whiteboardCalloutStorageConfigCodegen,
-} from '../callout/storage/callout-storage-config.params.request';
+} from '../../callout/storage/callout-storage-config.params.request';
 import {
   SpacePrivacyMode,
   SpaceVisibility,
@@ -43,15 +43,15 @@ import {
 import {
   createWhiteboardCollectionCalloutCodegen,
   createWhiteboardOnCalloutCodegen,
-} from '../callout/call-for-whiteboards/whiteboard-collection-callout.params.request';
+} from '../../callout/call-for-whiteboards/whiteboard-collection-callout.params.request';
 import {
   createPostCollectionCalloutCodegen,
   createPostCardOnCalloutCodegen,
-} from '../callout/post/post-collection-callout.params.request';
-import { createWhiteboardCalloutCodegen } from '../callout/whiteboard/whiteboard-callout.params.request';
-import { createReferenceOnProfileCodegen } from '../references/references.request.params';
-import { entitiesId } from '../roles/community/communications-helper';
-import { updateAccountPlatformSettingsCodegen } from '../account/account.params.request';
+} from '../../callout/post/post-collection-callout.params.request';
+import { createWhiteboardCalloutCodegen } from '../../callout/whiteboard/whiteboard-callout.params.request';
+import { createReferenceOnProfileCodegen } from '../../references/references.request.params';
+import { entitiesId } from '../../roles/community/communications-helper';
+import { updateAccountPlatformSettingsCodegen } from '../../account/account.params.request';
 
 const organizationName = 'org-name' + uniqueId;
 const hostNameId = 'org-nameid' + uniqueId;
