+++ test/functional-api/storage/files-to-upload/file-avif.avif
