+++ test/functional-api/storage/files-to-upload/file-gif.gif
