+++ test/functional-api/storage/files-to-upload/file-jpeg.jpeg
