+++ test/functional-api/storage/files-to-upload/file-jpg.jpg
