+++ test/functional-api/storage/files-to-upload/file-png.png
