+++ test/functional-api/storage/files-to-upload/file-webp.webp
