+++ test/functional-api/storage/files-to-upload/image.png
