+++ test/functional-api/storage/files-to-upload/vert.jpg
